# RavenDB-24100 Optimized search for the first free bit in bitmap.