@redknightlois 2025-05-21
LGTM. Having said that, there are a few improvements we can do on this. First, the benchmarks are not considering the case of how far from the beginning of the bitmap is the first zero; that would most likely give very different profiles. And because I am already imagining what that is going to show, we could technically improve it by loading a whole SIMD vector and check if it is full of ones. On large bitmaps that would be a great speedup.